refactor: saparate Search & Result
