+++ src/app/App.tsx
@@ -1,15 +1,15 @@
 import React from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import Search from './components/Search';
 import '../app/styles/style.css';
 import Page404 from './page/Page404';
+import Main from './page/Main';
 
 const App: React.FC = () => {
   return (
     <Router>
       <main>
         <Routes>
-          <Route path="/" element={<Search />} />
+          <Route path="/" element={<Main />} />
           <Route path="*" element={<Page404 />} />
         </Routes>
       </main>
